Split solution submission into two files
One file contains StableXSolutionSubmitting and the other the
retry_with_gas_price_increase because the single file got too big.

No logic has been changed. The relevant code has been moved into the new
file and imports adjusted.

In another commit I will introduce a trait for the retry function so
that StableXSolutionSubmitting can be tested independently.
